semimajor refactor :)
