+++ iframe/ipc.js
@@ -2,109 +2,29 @@ import { BitField } from './bitfield.js'
 import { DataReader, jsonToType } from '../data.js'
 import { codes } from './incomingPacket.js'
 import { frame } from './index.js'
-import { cbs, mouseMoveCb, pauseCb, wheelCb } from './controls.js'
-import { blockBreak, punchParticles, stepParticles } from './particles.js'
-import './world.js'
+import { buttons, options, listen, _cbs, _mouseMoveCb, _pauseCb, _wheelCb, _optionListeners } from 'api'
+import { Blocks, Items, Entities, BlockIDs, ItemIDs, EntityIDs, Block, Item, Entity } from 'definitions'
+import 'world'
+import { _setPaused } from './api.js'
 
-Block = class Block{
-	static placeSounds = []; static stepSounds = []
-	static solid = true
-	static texture = null
-	static climbable = false
-	static gooeyness = 0
-	static breaktime = 3
-	place(x, y){
-		if(this.placeSounds.length)
-			sound(this.placeSounds[Math.floor(Math.random() * this.placeSounds.length)], x, y, 1, 0.8)
-	}
-	break(x, y){
-		if(this.placeSounds.length)
-			sound(this.placeSounds[Math.floor(Math.random() * this.placeSounds.length)], x, y, 1, 0.8)
-		blockBreak(this, x, y)
-	}
-	punch(x, y){
-		if(this.stepSounds.length)
-			sound(this.stepSounds[Math.floor(Math.random() * this.stepSounds.length)], x, y, 0.1375, 0.5)
-		punchParticles(this, x, y)
-	}
-	walk(x, y, e){
-		if(!e.alive) return
-		if(this.stepSounds.length)
-			sound(this.stepSounds[Math.floor(Math.random() * this.stepSounds.length)], x, y, 0.15, 1)
-		if((e.state & 0x1010000) == 0x10000 || (e.state & 4)) stepParticles(this, e)
-	}
-	fall(x, y){
-		if(this.stepSounds.length)
-			sound(this.stepSounds[Math.floor(Math.random() * this.stepSounds.length)], x ,y, 0.5, 0.75)
-	}
-}
-Item = class Item{
-	constructor(a){ this.count=a&255; this.name='' }
-	breaktime(b){ return b.breaktime }
-	static maxStack = 64
-	static model = 0
-}
-Entity = class Entity{
-	constructor(x,y){
-		this.ix = this.x = x
-		this.iy = this.y = y
-		this.dx = this.dy = 0
-		this.state = 0
-		this.age = 0
-		this.f = PI / 2
-		this.blocksWalked = 0
-	}
-	prestep(){
-		const { gooeyness } = getblock(floor(this.x), floor(this.dy > 0 ? this.y : this.y + this.height / 4))
-		if(gooeyness) this.dx *= (1 - gooeyness), this.dy *= (1 - gooeyness)
-	}
-	step(){
-		this.blocksWalked += abs(this.dx * dt)
-		if((this.state & 0x10000) && !(this.state & 2)){
-			if(this.blocksWalked >= 1.7){
-				this.blocksWalked = 0
-				const x = floor(this.x + this.dx * dt), y = ceil(this.y) - 1
-				const block = getblock(x, y)
-				if(block.walk) block.walk(x, y, this)
-			}
-		}else this.blocksWalked = this.dy < -10 ? 1.7 : 1.68
-	}
-	tick(){}
-	event(i){}
-	sound(a,b=1,c=1){sound(a, this.ix-.5, this.iy-.5+this.head, b, c)}
-	static width = 0.5
-	static height = 1
-	static head = .5
-	static gx = 1
-	static gy = 1
-	static alive = false
-}
-
-
-buttons = new BitField()
-options = {}
 loaded = () => {
 	for(const data of msgQueue)onMsg({data})
 	msgQueue = null
 }
-const optionListeners = {}
-listen = (...keys) => {
-	const cb = keys.pop()
-	for(const key of keys){
-		(optionListeners[key] || (optionListeners[key] = [])).push(cb)
-		if(key in options) cb(options[key])
-	}
-}
 listen('music', () => _bgGain.gain.value = options.music * options.music)
+listen('sound', () => _volume = options.sound)
 
 const onMsg = ({data}) => {
 	if(Array.isArray(data)){
 		if(data.length == 2){
-			if(typeof data[0] == 'string') options[data[0]] = data[1]
-			else for(const cb of mouseMoveCb) cb(data[0], data[1])
+			const [a, b] = data
+			if(typeof a == 'string'){
+				options[a] = b
+				if(_optionListeners[a]) for(const f of _optionListeners[a]) f(b)
+			}else for(const cb of _mouseMoveCb) cb(a, b)
 			return
 		}else if(data.length == 1){
-			for(const cb of wheelCb) cb(data[0])
+			for(const cb of _wheelCb) cb(data[0])
 			return
 		}
 		// import scripts
@@ -166,17 +86,16 @@ const onMsg = ({data}) => {
 	}else if(typeof data == 'number'){
 		if(data >= 0){
 			buttons.set(data)
-			if(cbs[data])for(const f of cbs[data])f()
+			if(_cbs[data])for(const f of _cbs[data])f()
 		}else buttons.unset(~data)
 	}else if(typeof data == 'boolean'){
-		paused = data
-		for(const cb of pauseCb) cb()
+		_setPaused(data)
+		for(const cb of _pauseCb) cb()
 	}
 }
 for(const data of [msgQueue, msgQueue = []][0]) onMsg({data})
 addEventListener('message', onMsg)
 onmessage = null
 
-pause = paused => me && postMessage(!!paused, '*')
 send = buf => postMessage(buf.build ? buf.build().buffer : buf.buffer || buf, '*')
 download = blob => postMessage(blob, '*')
